create new api module
