+++ api/src/main/java/org/esbtools/message/admin/model/HeaderType.java
@@ -1,0 +1,5 @@
+package org.esbtools.message.admin.model;
+
+public enum HeaderType {
+    JMS,METADATA
+}
