+++ common/pom.xml
@@ -27,6 +27,11 @@
     <distributionManagement></distributionManagement>
     <dependencies>
         <dependency>
+            <groupId>org.esbtools.message.admin</groupId>
+            <artifactId>esb-message-admin-api</artifactId>
+            <version>0.1.4-SNAPSHOT</version>
+        </dependency>
+        <dependency>
             <groupId>commons-codec</groupId>
             <artifactId>commons-codec</artifactId>
         </dependency>
