+++ rest-save-client/pom.xml
@@ -22,7 +22,7 @@
         </dependency>
         <dependency>
             <groupId>org.esbtools.message.admin</groupId>
-            <artifactId>esb-message-admin-common</artifactId>
+            <artifactId>esb-message-admin-api</artifactId>
             <scope>provided</scope>
 	    <version>0.1.4-SNAPSHOT</version>
         </dependency>
