+++ rest-save/pom.xml
@@ -14,7 +14,7 @@
     <dependencies>
         <dependency>
             <groupId>org.esbtools.message.admin</groupId>
-            <artifactId>esb-message-admin-common</artifactId>
+            <artifactId>esb-message-admin-api</artifactId>
             <version>0.1.4-SNAPSHOT</version>
         </dependency>
         <dependency>
